requirements.txt
fastapi
uvicorn
pandas
openpyxl
google-cloud-storage
